Struct RustSampler::sampler_engine::SamplerEngine

pub struct SamplerEngine {
    num_voices: u8,
    sound_bank: HashMap<u8, (String, f32, RingBuffer<f32>, SamplerVoice)>,
    file_names: Vec<String>,
    warp_buffer: RingBuffer<f32>,
    sampler_mode: SamplerMode,
    warp_voices: Vec<SamplerVoice>,
    sample_rate: f32,
    num_channels: usize,
    warp_sr_scalar: f32,
}

Fields
num_voices: u8sound_bank: HashMap<u8, (String, f32, RingBuffer<f32>, SamplerVoice)>file_names: Vec<String>warp_buffer: RingBuffer<f32>sampler_mode: SamplerModewarp_voices: Vec<SamplerVoice>sample_rate: f32num_channels: usizewarp_sr_scalar: f32
Implementations

impl SamplerEngine

pub fn new(sample_rate_: f32, num_channels_: usize) -> Self

pub fn process(&mut self) -> f32

pub fn add_to_paths_and_load(&mut self, file_path: &str)

Add a file to the paths of files saved in the file names and load file into the warp buffer.

pub fn add_file_to_paths(&mut self, file_path: &str)

Add a file to the paths of files saved in the file names.

pub fn load_file_from_path(&mut self, file_path: &str)

Load file from path into the warp buffer without loading into the file names.

pub fn assign_file_to_midi(&mut self, file_path: &str, note: u8)

Assigns an audio file to a midi note for the sound bank. (Assign mode)

Will add file to paths if not already there

pub fn note_on(&mut self, note: u8, velocity: f32)

Triggers a “note on” message and allocates a voice, stealing if necessary

pub fn note_off(&mut self, note: u8)

Triggers a note off message

pub fn set_adsr_warp( &mut self, attack_: f32, decay_: f32, sustain_: f32, release_: f32 )

Sets the attack, decay, sustain, and release for all the warp sample voices

pub fn set_adsr_assign( &mut self, attack_: f32, decay_: f32, sustain_: f32, release_: f32, note_of_assigned: u8 )

Sets the attack, decay, sustain, and release for the given assigned note

pub fn get_adsr_warp(&mut self) -> (f32, f32, f32, f32)

Returns attack, decay, sustain, release values for the warping sampler

Returns tuple in format: (attack,decay,sustain,release)

pub fn get_adsr_assign(&mut self, note_of_assigned: u8) -> (f32, f32, f32, f32)

Returns attack, decay, sustain, release values for the given assigned note

Returns tuple in format: (attack,decay,sustain,release)

pub fn set_num_voices(&mut self, num_voices: u8)

Sets the max number of voices in the warp sampler

pub fn get_num_voices(&mut self) -> u8

Returns the number of voices available for the warping sampler

pub fn set_mode(&mut self, mode: SamplerMode)

Sets the sampler mode (Warp, Assign, Sfz)

pub fn set_warp_base(&mut self, base_note: u8)

Sets the note for the warping to be based on

pub fn set_points_warp(&mut self, start_point: f32, end_point: f32)

Sets the start and end points for each of the voices for the warping sampler

start_point: (0%-100%), end_point: (0%-100%)

If the start point is greater than the endpoint, the playback will be reversed

pub fn get_points_warp(&mut self) -> (f32, f32)

Gets the start and end points (in percent) for the warp sampler

Returns tuple in the format: (start_point, end_point)

pub fn set_points_assign( &mut self, start_point: f32, end_point: f32, note_of_assigned: u8 )

Sets the start and end points for an assigned sampler voice

start_point: (0%-100%), end_point: (0%-100%), note_of_assignment: note of the

If the start point is greater than the endpoint, the playback will be reversed

pub fn get_points_assign(&mut self, note_of_assigned: u8) -> (f32, f32)

Gets the start and end points (in percent) of the voice assigned to the given midi note

Returns tuple in the format: (start_point, end_point)

pub fn set_sus_points_warp(&mut self, start_point: f32, end_point: f32)

Sets the start and end points of the warp buffer’s sustain looping. Values will be clamped within start and end points of the sample as a whole

pub fn get_sus_points_warp(&mut self) -> (f32, f32)

Gets the start and end points for the sustain loop of the warp sampler.

Returns tuple in the format: (start_point, end_point)

pub fn set_sus_points_assign( &mut self, start_point: f32, end_point: f32, note_of_assigned: u8 )

Sets the start and end points of the assigned buffer’s sustain looping. Values will be clamped within start and end points of the sample as a whole

pub fn get_sus_points_assign(&mut self, note_of_assigned: u8) -> (f32, f32)

Gets the start and end points for the sustain loop of the assigned note.

Returns tuple in the format: (start_point, end_point)

pub fn set_sus_looping_warp(&mut self, mode: SustainModes)

pub fn set_sus_looping_assign( &mut self, mode: SustainModes, note_of_assigned: u8 )

pub fn set_fade_time_warp(&mut self, fade_time: f32)

Sets crossfade time in seconds for the warp sampler, expects values between (0.00001 and 0.1)

pub fn set_fade_time_assign(&mut self, fade_time: f32, note_of_assigned: u8)

Sets crossfade time in seconds for the selected file, expects values between (0.00001 and 0.1)

fn get_voice_id(&mut self) -> usize

Chooses a voice and steals the quietest one